Just some minor code updates (#17)
Adding modifier
removing unused imports
general codecleanup

+++ enderio-base/src/main/java/com/enderio/base/common/util/TeleportUtils.java
@@ -17,8 +17,8 @@ public class TeleportUtils {
         if (!entity.level.isClientSide) {
             for (int i = 0; i < 16; ++i) {
                 double d3 = entity.getX() + entity.getRandom().nextDouble() * range;
-                double d4 = Mth.clamp(entity.getY() + (entity.getRandom().nextDouble()) * range / 2.0D, (double) entity.level.getMinBuildHeight(),
-                    (double) (entity.level.getMinBuildHeight() + ((ServerLevel) entity.level).getLogicalHeight() - 1));
+                double d4 = Mth.clamp(entity.getY() + (entity.getRandom().nextDouble()) * range / 2.0D,entity.level.getMinBuildHeight(),
+                    entity.level.getMinBuildHeight() + ((ServerLevel) entity.level).getLogicalHeight() - 1);
                 double d5 = entity.getZ() + entity.getRandom().nextDouble() * range;
                 if (entity.isPassenger()) {
                     entity.stopRiding();
